# lib/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: api
  version: 0.1.0
  description: Orivellum local-first AI workspace API
servers:
  - url: /api
paths:
...
  /works:
    get:
      operationId: listWorks
      summary: List all Works
      tags: [works]
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: work_type
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Works list
          content:
            application/json:
              schema:
                type: object
                properties:
                  works: { type: array, items: { $ref: "#/components/schemas/Work" } }
    post:
      operationId: createWork
      summary: Create a new Work
      tags: [works]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkCreate" }
      responses:
        "200":
          description: Created Work
          content:
            application/json:
              schema:
                type: object
                properties:
                  work: { $ref: "#/components/schemas/Work" }
  /works/{workId}:
    get:
      operationId: getWork
      summary: Get a Work by ID
      tags: [works]
      parameters:
        - name: workId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Work
          content:
            application/json:
              schema:
                type: object
                properties:
                  work: { $ref: "#/components/schemas/Work" }
    patch:
      operationId: updateWork
      summary: Update a Work
      tags: [works]
      parameters:
        - name: workId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkUpdate" }
      responses:
        "200":
          description: Updated Work
          content:
            application/json:
              schema:
                type: object
                properties:
                  work: { $ref: "#/components/schemas/Work" }
    delete:
      operationId: deleteWork
      summary: Delete a Work
      tags: [works]
      parameters:
        - name: workId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
  /works/{workId}/documents:
    get:
      operationId: getWorkDocuments
      summary: List documents for a Work
      tags: [works]
      parameters:
        - name: workId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents: { type: array, items: { $ref: "#/components/schemas/Document" } }
                  count: { type: integer }
  /works/{workId}/knowledge:
    get:
      operationId: getWorkKnowledge
      summary: List knowledge for a Work
      tags: [works]
      parameters:
        - name: workId
          in: path
          required: true
          schema: { type: string }
        - name: kind
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Knowledge items
          content:
            application/json:
              schema:
                type: object
                properties:
                  knowledge: { type: array, items: { $ref: "#/components/schemas/KnowledgeItem" } }
                  count: { type: integer }
  /works/{workId}/tasks/{taskId}:
    patch:
      operationId: updateWorkTask
      summary: Update a task (status, text)
      tags: [works]
      parameters:
        - name: workId
          in: path
          required: true
          schema: { type: string }
        - name: taskId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskUpdate" }
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                type: object
                properties:
                  task: { $ref: "#/components/schemas/Task" }
  /works/{workId}/tasks:
    get:
      operationId: getWorkTasks
      summary: List tasks for a Work
      tags: [works]
      parameters:
        - name: workId
          in: path
          required: true
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks: { type: array, items: { $ref: "#/components/schemas/Task" } }
                  count: { type: integer }
    post:
      operationId: createWorkTask
      summary: Create a task for a Work
      tags: [works]
      parameters:
        - name: workId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskCreate" }
      responses:
        "200":
          description: Created task
          content:
            application/json:
              schema:
                type: object
                properties:
                  task: { $ref: "#/components/schemas/Task" }
  /works/{workId}/stats:
    get:
      operationId: getWorkStats
      summary: Get aggregated stats for a Work
      tags: [works]
      parameters:
        - name: workId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Work stats
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkStats" }
  /works/{workId}/conversations:
    get:
      operationId: getWorkConversations
      summary: List conversations for a Work
      tags: [works]
      parameters:
        - name: workId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversations: { type: array, items: { $ref: "#/components/schemas/Conversation" } }

  # ─── Conversations ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Work:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        work_type: { type: string }
        description: { type: string, nullable: true }
        status: { type: string }
        meta: { type: object }
        doc_count: { type: integer }
        knowledge_count: { type: integer }
        pending_tasks: { type: integer }
        created_at: { type: string }
        updated_at: { type: string }
...
    WorkCreate:
      type: object
      required: [title]
      properties:
        title: { type: string }
        work_type: { type: string, default: research }
        description: { type: string, nullable: true }
        meta: { type: object }
    WorkUpdate:
      type: object
      properties:
        title: { type: string, nullable: true }
        description: { type: string, nullable: true }
        status: { type: string, nullable: true }
        meta: { type: object, nullable: true }
    WorkStats:
      type: object
      properties:
        work_id: { type: string }
        documents_by_kind: { type: object }
        knowledge_by_kind: { type: object }
        tasks_by_status: { type: object }
    Document:
      type: object
      properties:
        id: { type: string }
        work_id: { type: string, nullable: true }
        title: { type: string, nullable: true }
        source: { type: string, nullable: true }
        sha256: { type: string, nullable: true }
        kind: { type: string, nullable: true }
        classification: { type: string, nullable: true }
        readiness: { type: string }
        word_count: { type: integer, nullable: true }
        error_message: { type: string, nullable: true }
        meta: { type: object }
        created_at: { type: string }
...
    KnowledgeItem:
      type: object
      properties:
        id: { type: string }
        work_id: { type: string, nullable: true }
        kind: { type: string }
        text: { type: string }
        subject: { type: string, nullable: true }
        predicate: { type: string, nullable: true }
        object: { type: string, nullable: true }
        confidence: { type: number }
        review_status: { type: string }
        created_at: { type: string }
    Task:
      type: object
      properties:
        id: { type: string }
        work_id: { type: string, nullable: true }
        text: { type: string }
        status: { type: string }
        priority: { type: integer }
        created_at: { type: string }
    TaskUpdate:
      type: object
      properties:
        status: { type: string, nullable: true }
        text: { type: string, nullable: true }
    TaskCreate:
      type: object
      required: [text]
      properties:
        text: { type: string }
        priority: { type: integer, default: 0 }
    Conversation:
      type: object
      properties:
        id: { type: string }
        work_id: { type: string, nullable: true }
        title: { type: string, nullable: true }
        archived: { type: integer }
        model: { type: string, nullable: true }
        last_message: { type: string, nullable: true }
        message_count: { type: integer }
        created_at: { type: string }
        updated_at: { type: string }
...
    OkResponse:
      type: object
      properties:
        ok: { type: boolean }